Squashed commit of the following:
commit a21facc53140adf0034a01bc462a27c1de923670
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Mon Nov 27 18:45:48 2023 +0100

    wip

commit 76c9b92a260c969e14e92471e10e6ea2b531d7c0
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Fri Nov 24 19:36:04 2023 +0100

    wip / change to Bits

commit 9965db98faf03c1490eb2a31f177f61e9cd49def
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Fri Nov 24 18:12:40 2023 +0100

    wip

commit dd3fe1b317a88c6cae1e066bc43b32ce2d666f76
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Thu Nov 23 17:46:55 2023 +0100

    wip

commit f9ff953c9f5a06d951fbf14c496c475702a22e1b
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Wed Nov 22 19:44:21 2023 +0100

    wip

commit b03c558478b203106ad8d7764bae5abddf57fbc8
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Wed Nov 22 18:03:10 2023 +0100

    wip

commit 93a65f224202ed10c966ca3053c2817fdb6272bd
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Wed Nov 22 15:00:55 2023 +0100

    wip

commit 7773b1462682c621f668493aa567285b3d471675
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Tue Nov 21 18:26:05 2023 +0100

    wip

commit 40989924e32c6b1fa7fd9a3c9edca27911303b38
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Tue Nov 21 18:04:41 2023 +0100

    wip

commit 71a914dda96b6dd843fbadf0470a1d90f34dce6f
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Mon Nov 20 12:34:55 2023 +0100

    wip

commit aba27a4cc2324683ebff92b0c6eb857f1b2d66d0
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Sat Nov 18 19:04:56 2023 +0100

    wip

commit 4e381e79c4f9e620396c4213e32cb4144569022a
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Sat Nov 18 13:35:41 2023 +0100

    wip

commit 203461dcba1797aef6a3b17608cbcc49980d5e74
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Fri Nov 17 15:44:57 2023 +0100

    wip

commit 241e254abb43277d5c3475c0a4a972ccad61900d
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Fri Nov 17 00:23:59 2023 +0100

    wip

commit a772d94a398a7ad11c46760847b4e1b3066b43a1
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Thu Nov 16 15:10:56 2023 +0100

    wip

commit 895d321f9b68e85ab35a51b7466c0144ad590dfb
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Wed Nov 15 13:49:05 2023 +0100

    wip

commit 7f190f1daf0bc119a257b0f1293bde70a917b748
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Wed Nov 15 10:15:20 2023 +0100

    rollback of usage/refcount

commit 3c5204f2851a4c09978cc8beb6ab9d0b5dfbcd37
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Tue Nov 14 18:58:16 2023 +0100

    wip

commit 16a86878b62b802cfc0cf0ffa6ed14ffe65d3183
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Thu Nov 9 14:19:21 2023 +0100

    wip: prefetch

commit f75b3ee84efdcf852ac34b31c1049e155ee3beb7
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Thu Nov 9 13:46:40 2023 +0100

    wip: TileReader working

commit 52f1d37feeb5119ce10b9710f2115a16a89869d7
Author: clarisma <56932851+clarisma@users.noreply.github.com>
Date:   Wed Nov 8 18:51:06 2023 +0100

    wip

+++ src/geom/Box.h
@@ -199,7 +199,9 @@ public:
 
 	Coordinate center() const
 	{
-		return Coordinate((minX() + maxX()) / 2, (minY() + maxY()) / 2);
+		return Coordinate(
+			static_cast<int32_t>((static_cast<int64_t>(minX()) + maxX()) / 2),
+			static_cast<int32_t>((static_cast<int64_t>(minY()) + maxY()) / 2));
 	}
 
 	int minX() const { return m_minX; }
